HH-80408 do not retry post requests
