+++ client-tests/src/test/java/ru/hh/jclient/common/AdaptiveBalancingClientTest.java
@@ -1,0 +1,9 @@
+package ru.hh.jclient.common;
+
+public class AdaptiveBalancingClientTest extends BalancingClientTestBase {
+
+  @Override
+  protected boolean isAdaptive() {
+    return true;
+  }
+}
