+++ client/src/main/java/ru/hh/jclient/common/balancing/ResponseTimeTracker.java
@@ -8,7 +8,7 @@ public class ResponseTimeTracker {
   private volatile boolean isWarmUp = true;
   private int current;
 
-  public ResponseTimeTracker(int n) {
+  ResponseTimeTracker(int n) {
     this.n = n;
     times = new long[n];
   }
